Moved to an event-based loop
This commit introduces an event-based loop based on epoll. A main loop
handles the signal and ring-buffer handling. This change prepares the
program for handling Netlink messages and timer events in the event loop
for more efficient handling and better readability.

The event loop also processes all events by their priority in the following order:
1. Signals
2. Netlink (Not implemented in this commit)
2. Ring buffer

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>