Adds epoll event handling and caching
This git commit changes the neighsnoopd to become fully event-driven
using epoll and fully cached. The new event system is still
single-threaded; however, the events that epoll handles are as follows:

- We monitor Netlink for all kernel RTMGRP_IPV4_IFADDR,
  RTMGRP_IPV6_IFADDR, RTMGRP_LINK, RTMGRP_NEIGH events. This change
allows neighsnoopd to track and cache changes instead of doing lookups.

- We cache all relevant information about links, the relevant subnets,
  the bridge FDB, and the neighbor entries on the subnets we monitor.

- All eBPF ringbuffer events are epoll events. We also use the LPM TRIE
  data structure to filter out ARP/NS that are not destined for our
target subnets.

- Signal handling is now handled as events. A future goal for signal
  handling is to trigger functionality, such as forcing rereading all
the Netlink entries or rereading a potential config file.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>