Adds a stats client/server architecture
This commit introduces a trivial Unix socket-based client program that
shows the neighsnoopd cache's status in JSON by echoing the response
directly from the neighsnoopd. For now, this is used for debugging, but in
the future, it could be a venue for more useful information.

Some of the limitations of the client/server architecture are:
- The server only allows one client connection at a time. This means that
  if the client is stuck in a pipe or suspended without closing the socket,
it will block all other clients. Fixing this in the future is trivial.
- The client is a dumb echo server for now.
- There is no way to pass a parameter to the server to limit the caching
  information sent to the client in this commit.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>