Added timer support and neigh triggers
This commit introduces a timer using the timefd file descriptor, which
we monitor with epoll. It tracks target neighbors in reachable or stale
states and adds a timer to send a gratuitous ARP/NS.

The formula we use to calculate when we will send a gratuitous ARP/NS is
as follows:

brt_ms = base_reachable_time_ms
time = brt_ms / 4.0 / 1000.0 + (random % 2000) / 1000.0

The base_reachable_time_ms is gotten from the
/proc/sys/net/{ipv4,ipv6}/neigh/<ifname>/base_reachable_time_ms.

We chose the formula arbitrarily. The idea is that the gratuitous ARP/NS
should happen before the kernel changes the neighbor from reachable to
stale. Therefore, we divide the time by four to be sure. Then, to ensure
that we don't get all the gratuitous requests simultaneously, we add a
random time of up to two seconds.

We also updated the client/server to show the timer cache.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>